refactor: test코드들 정상화 - í 실제 api 작동 테스팅 - ë©api 작동 하지만, h2가 이전프로젝트 DB를 ã나타내는 문제
