--- backend/src/main/resources/data.sql
@@ -1,5 +1,0 @@
-insert into product (name, price, stock, status, created_at, modified_at)
-values ('Product A', 5000, 100, true, NOW(), NOW());
-
-insert into product (name, price, stock, status, created_at, modified_at)
-values ('Product B', 10000, 100, true, NOW(), NOW());
