+++ backend/src/test/java/productTest/ProductRepositoryTest.java
@@ -1,0 +1,4 @@
+package productTest;
+
+public class ProductRepositoryTest {
+}
